Move Js* types to be directly under Jint.Native namespace (#1745)
