+++ Jint/Native/Error/ErrorInstance.cs
@@ -3,13 +3,6 @@ using Jint.Runtime;
 
 namespace Jint.Native.Error;
 
-public sealed class JsError : ErrorInstance
-{
-    internal JsError(Engine engine) : base(engine, ObjectClass.Error)
-    {
-    }
-}
-
 public class ErrorInstance : ObjectInstance
 {
     private protected ErrorInstance(Engine engine, ObjectClass objectClass)
