+++ Jint/Native/Function/Function.cs
@@ -2,7 +2,6 @@ using System.Diagnostics;
 using System.Runtime.CompilerServices;
 using Esprima.Ast;
 using Jint.Native.Object;
-using Jint.Native.Proxy;
 using Jint.Runtime;
 using Jint.Runtime.Descriptors;
 using Jint.Runtime.Environments;
