+++ Jint/Native/Object/ObjectPrototype.cs
@@ -2,7 +2,6 @@
 
 using Jint.Collections;
 using Jint.Native.Array;
-using Jint.Native.Proxy;
 using Jint.Native.Symbol;
 using Jint.Runtime;
 using Jint.Runtime.Descriptors;
