+++ Jint/Native/SharedArrayBuffer/SharedArrayBufferConstructor.cs
@@ -1,9 +1,7 @@
 using Jint.Collections;
-using Jint.Native.DataView;
 using Jint.Native.Function;
 using Jint.Native.Object;
 using Jint.Native.Symbol;
-using Jint.Native.TypedArray;
 using Jint.Runtime;
 using Jint.Runtime.Descriptors;
 using Jint.Runtime.Interop;
