+++ Jint/Runtime/Interpreter/Expressions/JintIdentifierExpression.cs
@@ -2,7 +2,6 @@ using System.Diagnostics.CodeAnalysis;
 using System.Runtime.CompilerServices;
 using Esprima.Ast;
 using Jint.Native;
-using Jint.Native.Argument;
 using Jint.Runtime.Environments;
 using Environment = Jint.Runtime.Environments.Environment;
